requirements.txt
pandas>=1.5.0
numpy>=1.23.0
matplotlib>=3.6.0
seaborn>=0.12.0
scikit-learn>=1.2.0
xgboost>=1.7.0
tensorflow>=2.15.0
keras>=3.0.0
requests>=2.28.0
streamlit>=1.28.0
plotly>=5.14.0
jupyter>=1.0.0
notebook>=6.5.0
pymongo>=4.6.0
python-dotenv>=1.0.0
dnspython>=2.4.0

# FastAPI Backend
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# SHAP Explanations
shap>=0.44.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0

# Code Quality
black>=23.0.0
flake8>=6.1.0
isort>=5.12.0
mypy>=1.7.0
